refactor(outputView): LineStringFormatter를 OutputView로 통합
