+++ src/main/java/ladder/view/OutputView.java
@@ -6,6 +6,10 @@ import java.util.List;
 public class OutputView {
     private static final int PLAYER_NAME_PAD_SIZE = 6;
     private static final String RESULT_DESCRIPTION = "실행 결과";
+    private static final String VERTICAL_CHAR = "|";
+    private static final String HORIZONTAL_CHAR = "-";
+    private static final String EMPTY_SPACE = " ";
+    private static final int PATH_WIDTH = 5;
 
     private OutputView() {
     }
@@ -25,12 +29,27 @@ public class OutputView {
     public static void printLadder(List<LineDto> ladder) {
         StringBuilder sb = new StringBuilder();
         for (LineDto ld : ladder) {
-            sb.append(LineStringFormatter.create(ld));
+            sb.append(OutputView.createLineString(ld));
         }
         System.out.println(sb);
     }
 
     private static String padName(String name) {
         return String.format("%-" + PLAYER_NAME_PAD_SIZE + "s", name);
+    }
+
+    public static String createLineString(LineDto lineDto) {
+        StringBuilder sb = new StringBuilder();
+        lineDto.getConnected()
+                .forEach(val -> sb.append(convertLine(val)));
+        sb.append(VERTICAL_CHAR + "\n");
+        return sb.toString();
+    }
+
+    private static String convertLine(boolean connected) {
+        if (connected) {
+            return VERTICAL_CHAR + HORIZONTAL_CHAR.repeat(PATH_WIDTH);
+        }
+        return VERTICAL_CHAR + EMPTY_SPACE.repeat(PATH_WIDTH);
     }
 }
